refactor: add project service to fetch preview images
