+++ src/app/home-page/projects/project-item/project.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing'
+
+import { ProjectService } from './project.service'
+
+describe('ProjectService', () => {
+  let service: ProjectService
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({})
+    service = TestBed.inject(ProjectService)
+  })
+
+  it('should be created', () => {
+    expect(service).toBeTruthy()
+  })
+})
